+++ crates/full-node/sov-stf-runner/tests/integration/helpers/runner_init.rs
@@ -377,6 +377,7 @@ pub fn rollup_config_with_da<Da: DaService<Config = MockDaConfig>>(
         },
         sequencer: SequencerConfig {
             automatic_batch_production: true,
+            max_allowed_node_distance_behind: 10,
             // Set ttl to zero to disable for testing. This prevents nondeterminism.
             dropped_tx_ttl_secs: 0,
             admin_addresses: vec![],